# src/interfaces/http/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Backend task",
    "version": "v1"
  },
  "servers": [
    {
      "description": "Local server",
      "url": "/api"
    }
  ],
  "paths": {
    "/survey": {
      "get": {
        "operationId": "getAllsurvey",
        "tags": [ "Survey" ],
        "responses": {
          "200": {
            "description": "List of all survey submitted by user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Survey"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createSurvey",
        "tags": [ "Survey" ],
        "requestBody": {
          "description": "Survey data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewSurvey"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Survey created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Survey"
                }
              }
            }
          },
          "400": {
            "description": "Survey not created because of validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Survey": {
        "allOf": [
          { "$ref": "#/components/schemas/NewSurvey" },
          {
            "required": [ "id" ],
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ]
      },
      "NewSurvey": {
        "required": [ "question", "option" ],
        "properties": {
          "question": {
            "type": "string"
          },
          "option": {
            "type": "boolean"
          }
        }
      },
      "ValidationError": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [ "ValidationError" ]
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDetail"
            }
          }
        }
      },
      "ValidationErrorDetail": {
        "properties": {
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "NotFoundError": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [ "NotFoundError" ]
          },
          "details": {
            "type": "string",
            "enum": [ "Survey with id {id} not found" ]
          }
        }
      }
    }
  }
}